Moved body result to variable.
Seems like the body result is in a stream so you can only read it once (after that you will probably be at the end of the stream). So trying to read twice for access_token and expires_in generated an error (not found on the second read). Moving the Body to a JObject variable solved the issue.