Aligning cache duration to token duration (#38)
* Update Backend OAuth2 Authentication With Cache.policy.xml

Remove token cache duration name value parameter and derive the value from the token received from token provider. Token duration and cache duration are now linked.

* Update Backend OAuth2 Authentication With Cache.policy.xml

Fixing typo.

* Update Backend OAuth2 Authentication With Cache.policy.xml

Strong typing tokenDurationSeconds

* Moved body result to variable.

Seems like the body result is in a stream so you can only read it once (after that you will probably be at the end of the stream). So trying to read twice for access_token and expires_in generated an error (not found on the second read). Moving the Body to a JObject variable solved the issue.